Ported changes from ZDoom revision 966-971 / GZDoom revision 105: - Fixed: Parsing sector special bit masks must be done backwards so that later definitions take precedence. - Added base translation tables for UDMF compatibility maps which only should handle the native line and sector types of each game. - Turned the inactive SILENT_INSTANT_FLOORS define into a compatibility option so that it can be (un)set in a map definition and the menu. - Added the C99 printf size specifiers 't' (ptrdiff_t) and 'z' (size_t) to FString::Format() so that I can fix all the problem printf strings that a 64-bit GCC compile finds. - Changed the types of object hash indices in FArchive from size_t to DWORD. This seems to fix crashes on GCC 64-bit builds when saving games. Not sure if it was a GCC bug or my bug, since it worked fine with VC++, but since the code that calculates the index only returns a DWORD, storing it as a size_t was rather pointless. - Changed: For actors not being spawned on the floor P_FindFloorCeiling should be used to prevent them from dropping through 3DMIDTEX textures. - Fixed: AMageStaffFX2::IsOkayToAttack overwrote the projectile's angle - Reverted fmodsound because it got converted to UTF-8 resulting in broken output for the copyright sign.
SVN r1346 (latestzdoom)

+++ src/r_things.cpp
@@ -539,7 +539,7 @@ void R_InitSkins (void)
 			strncpy (key, sc.String, sizeof(key)-1);
 			if (!sc.GetString() || sc.String[0] != '=')
 			{
-				Printf (PRINT_BOLD, "Bad format for skin %d: %s\n", i, key);
+				Printf (PRINT_BOLD, "Bad format for skin %d: %s\n", (int)i, key);
 				break;
 			}
 			sc.GetString ();
@@ -551,7 +551,7 @@ void R_InitSkins (void)
 				{
 					if (stricmp (skins[i].name, skins[j].name) == 0)
 					{
-						sprintf (skins[i].name, "skin%d", i);
+						sprintf (skins[i].name, "skin%d", (int)i);
 						Printf (PRINT_BOLD, "Skin %s duplicated as %s\n",
 							skins[j].name, skins[i].name);
 						break;
@@ -732,7 +732,7 @@ void R_InitSkins (void)
 		if (!remove)
 		{
 			if (skins[i].name[0] == 0)
-				sprintf (skins[i].name, "skin%d", i);
+				sprintf (skins[i].name, "skin%d", (int)i);
 
 			// Now collect the sprite frames for this skin. If the sprite name was not
 			// specified, use whatever immediately follows the specifier lump.
@@ -783,7 +783,7 @@ void R_InitSkins (void)
 
 				if (spr == 0 && maxframe <= 0)
 				{
-					Printf (PRINT_BOLD, "Skin %s (#%d) has no frames. Removing.\n", skins[i].name, i);
+					Printf (PRINT_BOLD, "Skin %s (#%d) has no frames. Removing.\n", skins[i].name, (int)i);
 					remove = true;
 					break;
 				}
